Agent Simple now updates solution report

+++ main.py
@@ -35,7 +35,7 @@ if __name__ == "__main__":
 
     #-- Define default arguments, in case you run the module as a script --#
     DEFAULT_STUDENT_GUID = 'template'
-    DEFAULT_AGENT_NAME   = 'Realistic' #HINT: Currently choose between {Random,Simple, Realistic}
+    DEFAULT_AGENT_NAME   = 'Simple' #HINT: Currently choose between {Random,Simple, Realistic}
     DEFAULT_MALMO_PATH   = os.environ['MALMO_ROOT'] # HINT: Change this to your own path
     DEFAULT_AIMA_PATH    = os.environ['AIMA_PATH']  # HINT: Change this to your own path, forward slash only, should be the 2.7 version from https://www.dropbox.com/s/vulnv2pkbv8q92u/aima-python_python_v27_r001.zip?dl=0) or for Python 3.x get it from https://github.com/aimacode/aima-python
     DEFAULT_MISSION_TYPE = 'small'  #HINT: Choose between {small,medium,large}
@@ -131,7 +131,12 @@ if __name__ == "__main__":
 
             print('Save the solution report to a specific file for later analysis and reporting...')
             fn_result = args.resultpath + 'solution_' + args.studentguid + '_' + agent_name + '_' +args.missiontype + '_' + str(i_training_seed) + '_' + str(i_rep)
-            foutput = open(fn_result+'.pkl', 'wb')
+            if os.path.exists(os.path.dirname(fn_result)):
+                foutput = open(fn_result+'.pkl', 'wb')
+            else:
+                os.makedirs(os.path.dirname(fn_result))
+                foutput = open(fn_result+'.pkl', 'wb')
+            
             pickle.dump(agent_to_be_evaluated.solution_report,foutput) # Save the solution information in a specific file, HiNT:  It can be loaded with pickle.load(output) with read permissions to the file
             foutput.close()
